\name{chopitsim}
\alias{chopitsim}
\title{Simulated Data for test chopit function}
\usage{data(chopitsim)}
\description{
Simulated data with two self-response questions (qself1,qself2) and
three vignettes (qvign1,qvign2,qvign3).  Random effect (RE) also included.

Survey respondents were asked in almost the same language for a
self-assessment and for an assessment of several hypothetical persons
described by written vignettes.  The vignettes for one particular domain
of political efficacy were the following:

qvign1: ``[Alison] lacks clean drinking water. She and her neighbors are
  supporting an opposition candidate in the forthcoming elections that
  has promised to address the issue.  It appears that so many people
  in her area feel the same way that the opposition candidate will
  defeat the incumbent representative.''
  
qvign2: ``[Jane] lacks clean drinking water because the government is
  pursuing an industrial development plan.  In the campaign for an
  upcoming election, an opposition party has promised to address the
  issue, but she feels it would be futile to vote for the opposition
  since the government is certain to win.''
  
qvign3: ``[Moses] lacks clean drinking water.  He would like to change
  this, but he can't vote, and feels that no one in the government
  cares about this issue.  So he suffers in silence, hoping something
  will be done in the future.''

The following question is then read to the respondent for each vignette
and for a self-assessment: How much say [does `name' / do you] have in
getting the government to address issues that interest [him / her /
you]?

For the self-assessment and each of the vignette questions, respondents
are given the same set of ordinal categories in which to respond, for
example ``(5) Unlimited say, (4) A lot of say, (3) Some say, (2) Little
say, (1) No say at all.''
}

\source{Based on R/sim/mc12rev/hopitmc.Ae.data/hopitmc.Ae.007.111.2000.1.dat}

\author{Jonathan Wand}

\keyword{datasets}
