\name{poleffna}
\alias{poleffna}
\title{Simulated Poltical Efficacy Data---with NAs, DEMO ONLY!}
\usage{data(poleff)}
\description{
  This is only used as an example on effect of missing data coding.
  Missing reponses are coded with NA here instead of zeroes.
}
\author{Jonathan Wand}
\keyword{datasets}
