\name{selfcare}
\alias{selfcare}
\title{Selfcare data for China}
\usage{data(selfcare)}
\description{

   Questions:

   1. Overall in the last 30 days, how much difficulty did [name of person/you] have with self-care, such as washing or dressing [yourself/himself/herself]?
   2. In the last 30 days, how much difficulty did [name of person/you] have in taking care of and maintaining [your/his/her] general appearance (e.g.grooming, looking neat and tidy etc.)

Response categories:

   1. None
   2. Mild
   3. Moderate
   4. Severe
   5. Extreme/Cannot Do

Vignettes:

   1. [Helena] pays a lot of attention to the way she looks. She requires no assistance with cleanliness, dressing and eating.***

   2. [Anne] takes twice as long as others to put on and take off clothes, but needs no help with this. Although it requires an effort, she is able to bathe and groom herself, though less frequently than before. She does not require help with feeding.***

   3. [Victor] usually requires no assistance with cleanliness, dressing and eating. He occasionally suffers from back pain and when this happens he needs help with bathing and dressing.***

   4. [Sandra] lives on her own and has no relatives or friends nearby. Because of her arthritis, she is house-bound. She often stays all day in the same clothes that she has slept in as changing clothes is too painful. A neighbour helps her wash herself.***

   5. [Sue] is quadriplegic and must be washed, groomed, dressed and fed by somebody else.***


   Variables:

   q2020: self-assessemnt

   q2101d: vignette 1

   q2119d: vignette 2

   q2113d: vignette 3

   q2105d: vignette 4

   q2117d: vignette 5

   q1008: married is '2'

   q1006: height

   q1001: sex

   q1002: age

   q1004: weight

   q1010: years of schooling.

}
\references{


  \emph{WHO's World Health Survey}
    by Lydia Bendib, Somnath Chatterji, Alena Petrakova, Ritu Sadana,
    Joshua A. Salomon, Margie Schneider, Bedirhan Ustun, Maria
    Villanueva

King, G., & Wand, J. (2007). Comparing Incomparable Survey Responses: Evaluating and Selecting Anchoring Vignettes. \emph{Political Analysis}, 15(1), 46-66. \doi{10.1093/pan/mpl011}

}

\keyword{datasets}
