\name{sleep}
\alias{sleep}
\title{Sleep data for China}
\usage{data(sleep)}
\description{

Question:

   1. In the last 30 days, how much difficulty do you think [name of person/you] had with sleeping, such as falling asleep, waking up frequently during the night or waking up too early in the morning?

   2. In the last 30 days, how much of a problem did you have due to not feeling rested and refreshed during the day?

Response categories:

   1. None
   2. Mild
   3. Moderate
   4. Severe
   5. Extreme/Cannot Do

Vignettes:

   1. [Mark] falls asleep every night within five minutes of going to bed. He sleeps soundly during the whole night and wakes up in the morning feeling well-rested.***

   2. [Paolo] has no trouble falling asleep at night and does not wake up during the night, but every morning he finds it difficult to wake up. He uses an alarm clock but falls back asleep after the alarm goes off. He is late to work on four out of five days.***

   3. [Noemi] falls asleep easily at night, but two nights a week she wakes up in the middle of the night and cannot go back to sleep for the rest of the night.***

   4. [Damien] wakes up almost once every hour during the night. When she wakes up in the night, it takes around 15 minutes for her to go back to sleep. In the morning she does not feel well-rested.***

   5. [Daniel] takes about two hours every night to fall asleep. He wakes up once or twice a night feeling panicked and takes more than one hour to fall asleep again. Three to four nights a week he wakes up in the middle of the night and cannot go back to sleep for the rest of the night.***


   q2080  : self-assessment

   q2119c : vignette 1

   q2103c : vignette 2

   q2107c : vignette 3

   q2109c : vignette 4

   q2115c : vignette 5

   q1008: married is '2'

   q1006: height

   q1001: sex

   q1002: age

   q1004: weight

   q1010: years of schooling.



}
\references{

  \emph{WHO's World Health Survey}
    by Lydia Bendib, Somnath Chatterji, Alena Petrakova, Ritu Sadana,
    Joshua A. Salomon, Margie Schneider, Bedirhan Ustun, Maria
    Villanueva

King, G., & Wand, J. (2007). Comparing Incomparable Survey Responses: Evaluating and Selecting Anchoring Vignettes. \emph{Political Analysis}, 15(1), 46-66. \doi{10.1093/pan/mpl011}

}

\keyword{datasets}
