\name{table1src}
\alias{table1src}
\title{Specific response values that have inequalities to create table1}
\usage{data(table1src)}
\description{

  Table with 13 observations.

  summary(anchors( y ~ z1 + z2 , data=table1src))

  will create contents of data(table1)

}
\author{Gary King and Jonathan Wand}
\references{
King, G., & Wand, J. (2007). Comparing Incomparable Survey Responses: Evaluating and Selecting Anchoring Vignettes. \emph{Political Analysis}, 15(1), 46-66. \doi{10.1093/pan/mpl011}
}

\keyword{datasets}
