% Created :  2002-10-05
% Modified:  $Date: 2004/10/17 20:09:46 $
% Revision:  $Revision: 1.10 $
% RCS-ID:    $Id: chopit.Rd,v 1.10 2004/10/17 20:09:46 jwand Exp $
\name{anchors.chopit.check}
\alias{anchors.chopit.check}
\title{Compound Hierarchical Ordered Probit (CHOPIT)}
\description{
Check internal consistency of options to be used by chopit.
}
\usage{
  anchors.chopit.check(count, options ) 
}
\arguments{
    \item{count}{object of class anchors.data.count.}
    \item{options}{object of class anchors.options.}
}
\details{

  This function is unlikely to be used directly by a user.
  
  Related materials and worked examples are available at
  http://wand.stanford.edu/anchors/
     
}

\author{Jonathan Wand \url{http://wand.stanford.edu}}

\references{
  Wand, Jonathan; Gary King; and Olivia Lau. (2007) ``Anchors: Software for
  Anchoring Vignettes''. \emph{Journal of Statistical Software}.  Forthcoming.
  copy at http://wand.stanford.edu/research/anchors-jss.pdf

  Wand, Jonathan and Gary King. (2007)
  Anchoring Vignetttes in R: A (different kind of) Vignette
  copy at http://wand.stanford.edu/anchors/doc/anchors.pdf

  King, Gary; Christopher J.L. Murray; Joshua A. Salomon; and Ajay
  Tandon. "Enhancing the Validity and Cross-cultural Comparability of
  Survey Research," American Political Science Review, Vol. 98, No. 1
  (February, 2004): 191-207,
  copy at http://gking.harvard.edu/files/abs/vign-abs.shtml
}
  
\seealso{\code{\link{anchors}},\code{\link{chopit}}}

\keyword{models}
