\name{table1}
\alias{table1}
\title{Results from cases given in King and Wand (2007)}
\usage{data(table1)}
\description{

  Table with 13 observations and two columns, "Cs" (indicating the lower
  bound of the range for the DIF-corrected variable), and "Ce"
  (indicating the upper case of the DIF-corrected variable).  This table
  provides a baseline for checking the output from the
  \code{\link{anchors}} example.  
}
\author{Gary King and Jonathan Wand}
\references{
  Gary King and Jonathan Wand.  "Comparing Incomparable Survey
  Responses: New Tools for Anchoring Vignettes," Political Analysis, 15,
  1 (Winter, 2007): Pp. 46-66,
  copy at http://gking.harvard.edu/files/abs/c-abs.shtml.
}

\keyword{datasets}
