\name{table1src}
\alias{table1src}
\title{Specific response values that have inequalities to create table1}
\usage{data(table1src)}
\description{

  Table with 13 observations.

  summary(anchors( y ~ z1 + z2 , data=table1src))

  will create contents of data(table1)
  
}
\author{Gary King and Jonathan Wand}
\references{
  Gary King and Jonathan Wand.  "Comparing Incomparable Survey
  Responses: New Tools for Anchoring Vignettes," Political Analysis, 15,
  1 (Winter, 2007): Pp. 46-66,
  copy at http://gking.harvard.edu/files/abs/c-abs.shtml.
}

\keyword{datasets}
